rename cache to access and remove unnecessary update to dictionary
